Viz optimizations (#84)
* Integrate virtual env for visualization

* Add basic plot example; update cmake execution in run_benchmark script

* Add pmem directory

* Rename viz folder

* Add parameter for help and config path

* Remove cmake options of server

* Exchange run_benchmark with visualize script

* Improve error handling

* Rename var

* Update visualize.sh

* Change assertion for args

* Add viz of raw jsons

* Add raw folder for testing

* Make visualize.sh script executable

* Add viz for multiple dimensions

* Rename json util classes

* Add viz for single dimension

* Remove results folder

* Remove unused variable

* Correct subheading order

* Remove results copy from plotter

* Integrate PR comments

* Print actual directories

* Allow multiple matrix jsons

* Add units to labels of continous args

* Add flag for deleting old png and html files

* Add duration boxplot for unidimensional matrix

* Remove typo

* Create variables for script args

* Check for non-existing matrix args

* Rename methods of PngCreator

* Change visualize.sh to setup_viz.sh

* Integrate pylint comments

- add module and class documentations
- improve import order
- remove unused vars
- use enumerate(x) instead of range(len(x))

* Add explanation of viz to README

* Require venv instead of virtualenv

* Complete README

* Add first simple fix.

* Add new categorical args.

* Fix bug.

* Fix plots for matrix_args=3

* Update README

+++ viz/main.py
@@ -1,37 +1,56 @@
-#! ../viz/venv/bin/python
-import os
-import argparse as ap
+"""
+This is the main module that has to be executed from the root directory, i.e. "perma-bench". It creates result PNGs and
+displays them on an user interface.
+"""
 
+# ! ../viz/venv/bin/python
+
+import argparse
+import os
+import sys
 import user_interface as ui
+
 from png_creator import PngCreator
 
 
 def dir_path(path):
+    """
+    Checks if the given directory path is valid.
+
+    :param path: directory path to the results folder
+    :return: bool representing if path was valid
+    """
     if os.path.isdir(path):
         return path
     else:
-        raise ap.ArgumentTypeError(f"The path to the results directory is not valid.")
+        raise argparse.ArgumentTypeError(f"The path to the results directory is not valid.")
 
 
 if __name__ == "__main__":
+    # check if python is running inside virtualenv or inside venv
+    if not (getattr(sys, 'real_prefix', None) or (getattr(sys, 'base_prefix', sys.prefix)) != sys.prefix):
+        sys.exit("Please run `source setup_viz.sh` once to setup the visualization environment.")
+
     # parse args + check for correctness and completeness of args
-    parser = ap.ArgumentParser()
+    parser = argparse.ArgumentParser()
     parser.add_argument("results_dir", type=dir_path, help="path to the results directory")
-    parser.add_argument("--delete", action="store_true", help="delete already existing png and html files")
+    parser.add_argument("--delete", action="store_true", help="delete already existing PNG and HTML files")
     args = parser.parse_args()
 
     # get directory paths of html and img folder
-    root_dir = os.path.abspath(os.curdir)
+    root_dir = os.path.abspath(os.getcwd())
     html_dir = os.path.join(root_dir, "viz/html")
     img_dir = os.path.join(os.path.abspath(args.results_dir), "img/")
 
     # delete already existing png and html files
     if args.delete:
-        for file in os.listdir(html_dir):
-            if file.endswith(".html"):
-                os.remove(os.path.join(html_dir, file))
-        for png in os.listdir(img_dir):
-            os.remove(os.path.join(img_dir, png))
+        if os.path.exists(html_dir):
+            for file in os.listdir(html_dir):
+                if file.endswith(".html"):
+                    os.remove(os.path.join(html_dir, file))
+        if os.path.exists(img_dir):
+            for png in os.listdir(img_dir):
+                os.remove(os.path.join(img_dir, png))
 
     # create img folder
     if not os.path.isdir(img_dir):
@@ -39,10 +58,11 @@ if __name__ == "__main__":
 
     # create pngs
     png_creator = PngCreator(args.results_dir, img_dir)
-    png_creator.create_pngs_for_raw_jsons()
-    png_creator.create_pngs_for_matrix_json()
+    png_creator.process_raw_jsons()
+    png_creator.process_matrix_jsons()
 
     # create user interface for pngs
+    print("Generating HTML pages")
     ui.init(img_dir)
-    print(f"The visualization is finished. The result png files are located in {img_dir} and can be viewed "
-          f"in a browser with \"<browser-name> {html_dir}/index.html\".")
+    print(f"\nThe visualization is finished. The result PNG files are located in {img_dir} and can be viewed in a "
+          f"browser with\n\t$ <browser-name> {html_dir}/index.html")
